Move `join_vars`
